Update the Content-Filtered Topic creation logic
Before this change, the conditional code for EPROSIMA correctly
used a distinct name for the CFT.  This change expands that fix
to apply to all implementations.

Also fixed a signed/unsigned warning -- I'm assuming sequence length()
is unsigned on all implementations since it's true for C++ vector
and for IDL-to-C++ mappings.